Fix: setting meta data for nmld files

+++ +ns/setMeta.m
@@ -1,9 +1,12 @@
-function setMeta(tbl,metaParm,metaVal,comments)
+function setMeta(tbl,metaParm,metaVal,comments,pv)
+% Function to set or add meta data to json files corresponding to
+% experiments, sessions, or subjects.
 arguments
     tbl
     metaParm (1,:) string
     metaVal (1,:) string
     comments (1,1) string  = ""
+    pv.dryrun (1,1) logical = true
 end
 
 % Sanity checks and singleton expansion
@@ -19,19 +22,24 @@ end
 
 % Loop over the table
 tplCntr = 0;
+if pv.dryrun
+    fprintf('DRYRUN:\n');
+end
+
 for tpl = fetch(tbl)'
     tplCntr = tplCntr +1;
     % Determine json file
     switch class(tbl)
         case "ns.Subject"
             error('niy')
-        case "ns.Experiment"
-            jsonFile = strrep(file(ns.Experiment &tpl),'.mat','.json');
+        case "ns.Experiment"            
+            [fldr,thisFile,~] = fileparts(file(ns.Experiment &tpl));            
         case "ns.Session"
             error('niy')
         otherwise
             error('Meta data for %s???',class(tbl));
     end
+    jsonFile =fullfile(fldr ,thisFile +".json");
     % Read or create metaData
     if exist(jsonFile,"file")
         metaData = readJson(jsonFile);
@@ -55,10 +63,15 @@ for tpl = fetch(tbl)'
     % Save changed meta data
     if changed
         % Write the updated JSON back to the file
-        fprintf('Update meta data for %s in %s\n',metaParm(tplCntr),jsonFile)
-        writeJson(metaData,jsonFile);
+        fprintf('Update meta data for %s in %s to %s\n',metaParm(tplCntr),jsonFile, metaVal(tplCntr))
+        if ~pv.dryrun              
+            writeJson(metaData,jsonFile);
+        end
     else
         fprintf('No change in meta data for %s in %s\n',metaParm(tplCntr),jsonFile)
     end
+end
+if pv.dryrun
+    fprintf('DRYRUN Complete.\n');
 end
 end
